+++ controllers/ConfigController.php
@@ -25,7 +25,7 @@ class ConfigController extends Controller
     /**
      * @inheritdoc
      */
-    public function getAccessRules()
+    protected function getAccessRules()
     {
         return [['permissions' => ManageModules::class]];
     }