+++ controllers/ListController.php
@@ -30,7 +30,10 @@ use yii\web\HttpException;
 
 class ListController extends AbstractTaskController
 {
-    public function getAccessRules()
+    /**
+     * @inheritdoc
+     */
+    protected function getAccessRules()
     {
         return [
             [ContentContainerControllerAccess::RULE_USER_GROUP_ONLY => [Space::USERGROUP_MEMBER, User::USERGROUP_SELF]],