+++ controllers/SearchController.php
@@ -25,7 +25,10 @@ use Yii;
 class SearchController extends AbstractTaskController
 {
 
-    public function getAccessRules()
+    /**
+     * @inheritdoc
+     */
+    protected function getAccessRules()
     {
         return [
             [ContentContainerControllerAccess::RULE_USER_GROUP_ONLY => [Space::USERGROUP_MEMBER, User::USERGROUP_SELF]]
@@ -67,4 +70,3 @@ class SearchController extends AbstractTaskController
     }
 
 }
-